re_grpc_server
Function serve

pub async fn serve(
    addr: SocketAddr,
    memory_limit: MemoryLimit,
    shutdown: Shutdown,
) -> Result<()>

Start a Rerun server, listening on addr.

A Rerun server is an in-memory implementation of a Storage Node.

The returned future must be polled for the server to make progress.

Currently, the only RPCs supported by the server are WriteMessages and ReadMessages.

Clients send data to the server via WriteMessages. Any sent messages will be stored in the server’s message queue. Messages are only removed if the server hits its configured memory limit.

Clients receive data from the server via ReadMessages. Upon establishing the stream, the server sends all messages stored in its message queue, and subscribes the client to the queue. Any messages sent to the server through WriteMessages will be proxied to the open ReadMessages stream.