re_grpc_server
Function spawn_with_recv

pub fn spawn_with_recv(
    addr: SocketAddr,
    memory_limit: MemoryLimit,
    shutdown: Shutdown,
) -> (Receiver<LogMsg>, Receiver<TableMsg>)

Start a Rerun server, listening on addr.

This function additionally creates a smart channel, and returns its receiving end. Any messages received by the server are sent through the channel. This is similar to creating a client and calling ReadMessages, but without the overhead of a localhost connection.

The server is spawned as a task on a tokio runtime. This function panics if the runtime is not available.

See serve for more information about what a Rerun server is.